anolis: erofs: fix the warning in erofs_fill_symlink()
ANBZ: #11420

Fixes the check_add_overflow() warning:
220 |      check_add_overflow(m_pofs, (unsigned int)inode->i_size, &off) ||
      |      ^~~~~~~~~~~~~~~~~~

Fixes: 6f206e39a95c ("erofs: fix incorrect symlink detection in fast symlink")
Link: https://gitee.com/anolis/cloud-kernel/pulls/4010

+++ fs/erofs/inode.c
@@ -217,7 +217,7 @@ static int erofs_fill_symlink(struct inode *inode, void *kaddr,
 	m_pofs += vi->xattr_isize;
 	/* check if it cannot be handled with fast symlink scheme */
 	if (vi->datalayout != EROFS_INODE_FLAT_INLINE || inode->i_size < 0 ||
-	    check_add_overflow(m_pofs, (unsigned int)inode->i_size, &off) ||
+	    check_add_overflow((loff_t)m_pofs, inode->i_size, &off) ||
 	    off > i_blocksize(inode)) {
 		inode->i_op = &erofs_symlink_iops;
 		return 0;
